+++ R/classify-diabetes.R
@@ -77,7 +77,7 @@ classify_diabetes <- function(
     diagnoser = diagnoser
   )
 
-  pregnancy_dates <- get_pregnancy_dates(
+  pregnancy_dates <- keep_pregnancy_dates(
     lpr2 = lpr2,
     lpr3 = lpr3
   )
@@ -86,7 +86,8 @@ classify_diabetes <- function(
   diabetes_diagnoses <- include_diabetes_diagnoses(
     lpr2 = lpr2,
     lpr3 = lpr3
-  )
+  ) |>
+    add_t1d_diagnoses_cols()
 
   podiatrist_services <- include_podiatrist_services(
     sysi = sysi,
@@ -95,8 +96,7 @@ classify_diabetes <- function(
 
   gld_purchases <- include_gld_purchases(
     lmdb = lmdb
-  ) |>
-    add_insulin_purchases_cols()
+  )
 
   hba1c_over_threshold <- include_hba1c(
     lab_forsker = lab_forsker
@@ -108,6 +108,11 @@ classify_diabetes <- function(
     exclude_pregnancy(
       pregnancy_dates = pregnancy_dates,
       included_hba1c = hba1c_over_threshold
+    ) |>
+    add_insulin_purchases_cols() |>
+    dplyr::select(
+      -"atc",
+      -"indication_code",
     )
 
   # Joining into an initial dataset -----
@@ -117,9 +122,18 @@ classify_diabetes <- function(
     gld_hba1c_after_exclusions = gld_hba1c_after_exclusions
   )
 
-  inclusions # |>
-  # create_inclusion_dates() |>
-  # classify_t1d()
+  inclusions |>
+    # create_inclusion_dates() |>
+    classify_t1d() |>
+    # If has_t1d is NA, t2d will also be NA
+    dplyr::mutate(has_t2d = !.data$has_t1d) |>
+    dplyr::select(
+      "pnr",
+      # "stable_inclusion_date",
+      # "raw_inclusion_date",
+      "has_t1d",
+      "has_t2d"
+    )
 }
 
 #' After inclusion and exclusion, classify those with type 1 diabetes.
@@ -131,10 +145,16 @@ classify_diabetes <- function(
 #' @keywords internal
 #'
 classify_t1d <- function(data) {
-  # data |>
-  #   get_has_t1d_primary_diagnosis() |>
-  #   get_only_insulin_purchases() |>
-  #   get_majority_of_t1d_primary_diagnosis() |>
-  #   get_insulin_purchases_within_180_days() |>
-  #   get_insulin_is_two_thirds_of_gld_purchases()
+  logic <- c(
+    "has_t1d"
+  ) |>
+    rlang::set_names() |>
+    purrr::map(get_algorithm_logic) |>
+    # To convert the string into an R expression
+    purrr::map(rlang::parse_expr)
+
+  data |>
+    dplyr::mutate(
+      has_t1d = !!logic$has_t1d
+    )
 }